fix(vip-618): revert mutation; host redeploy constants in vip-800
- Restore vip-618's original 10 buyback addresses, USDC_TO_SWEEP=14,986
  and U_MIN_OUT=7,418 — the VIP is queued on-chain and must match what
  governance audited.
- Define the redeployed buyback addresses (PR #162) and retuned
  single-multihop budget (USDC_TO_SWEEP=10,000, U_MIN_OUT=9,900 under
  the 9,996.60 U QuoterV2 read) locally in vip-800/part-1.
- BUYBACKS index order preserved so PSR-row → buyback mapping survives.
- vip-800/part-1 keeps re-exporting the untouched vip-618 constants so
  sims have a single import surface.
